[1020] Fixed dumb component
