+++ spec/components/blue_inset_text_with_link/view_preview.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module BlueInsetTextWithLink
+  class ViewPreview < ViewComponent::Preview
+    def default
+      render BlueInsetTextWithLink::View.new(title: "title", link_text: "link_text", url: "url")
+    end
+  end
+end
